python/requirements.txt
pandas
sklearn
numpy
pyserial